renamed the tables module
