+++ mindsdb/integrations/handlers/jira_handler/jira_tables.py
